// src/components/Button/Button.types.ts
export enum BUTTON_APPEARANCE {
  PRIMARY = 'primary',
  WARNING = 'warning',
  DANGER = 'danger',
  SUCCESS = 'success',
  SUBTLE = 'subtle',
  DEFAULT = 'default',
};

export enum BUTTON_SIZE {
  SMALL = 'sm',
  MEDIUM = 'md',
  LARGE = 'lg',
};

export enum BUTTON_SHAPE {
  ROUNDED = 'rounded',
  SQUARE = 'square',
  PILL = 'pill',
};